Partial add failing specs for the BottledWater client API
The Ruby adapter MUST implement the BottledWater client API in order
to be able to receive the messages from it.

 Overview of Bottled Water for PostgreSQL:

 _________  SQL connection       _______________________
|         |  (snapshot)         |           |           |
|   PG    |<············frame··>|           A   Ruby    |---Ruby
|  server |<-------frame------->| BW client P  adapter  |   hashes-->
|  with   |  Replication        |           I           |
| bottled | (stream of changes) |           |           |
|  water  |                     |  --(Avro messages)->  |
|extension|                     |___________|___________|
|_________|
                      Ruby client for Bottled Water for PostgreSQL

The BW client extracts the Avro encoded data from the frames and
exposes it to any adapter present. (The Kafka adapter would pass through
the Avro messages to Kafka, but the Ruby adapter will decode them
to build nested Ruby hashes.)

TODO: Define the actual API (see placeholders in the specs) and
specify which mecanisms will be used to transmit the `database_url`
from the Ruby adapter to the BottledWater client.