Add functions-based API to the client
Define accessors to set the client configuartion options and
the callbacks which will be called on any adapter (Kafka, Ruby...)

Provide functions to create a configuration context, and to free it.
Provide functions to start and stop the client (i.e. start waiting
for changes on either the SQL or replication connection, once changes
are available, the callbacks are called on the adapter).

Thank you @bgilbert!