Fix missing error handling when snapshot fails
When the client is started ti attempts to take a snapshot if
no replication slot exists. Part of taking a snapshot requires
to create the missing replication slot.

When the snapshot is successful, the replication slot is then used
to poll changes from the database. (If the client is restarted,
the presence of the replication slot will prevent a new snapshot
to be attempted.)

When the snapshot fails, however, the replication slot must be
dropped to allow new snapshot attempts.

Note: it may not always be possible to drop the replication slot,
in which case change polling will start even if no consistent
snapshot has been taken. In that case, manually dropping it, then
restarting the client fixes the issue.

Thank you @ept!