Add Client#register_insert_row_callback, with spec and example
This method is difficult to test directly because the 'insert_row'
callback function must be provided complex arguments, whose types
are not part of the Bottled Water API. In normal use, these arguments
are built by Bottled Water itself.

Taking that into account, I defined the register method in the same way
that I defined the #hello_ruby example, which can now be removed.

Because FFI callback functions must be defined as constants, which unusual
for Ruby methods, I decided to provide a default implementation as part
of the Ruby adapter API, and defined the Client::InsertRowCallback.

The specs verify that the Client#register_insert_row_callback actually
registers the correct callback using the corresponding function from
the Bottled Water API.
(The specs which verify that for that last function is attached properly
are temporary skipped for convenience, but that's a low-risk aspect of
the adapter.)

And... it works! :D